Revise Residents, add_edit_delete and Import_export

+++ app/Http/Controllers/FamilyMemberController.php
@@ -73,7 +73,7 @@ class FamilyMemberController extends Controller
             'gender'            => ['required', 'string', 'max:255', 'regex:/^[A-Za-z]+$/'],
             'birthdate'         => ['required', 'date_format:Y-m-d'],
             'address'           => ['required', 'string', 'max:255', 'alpha_spaces'],
-            'family_code'       => ['nullable', 'string', 'max:255'],
+            'family_code'       => ['nullable', 'string', 'max:255', 'exists:families,family_code'],
             'is_family_head'    => ['required', 'string', 'max:255'],
             'sectoral_classification' => ['required', 'string', 'max:255', 'alpha_spaces'],
         ]);
@@ -90,17 +90,18 @@ class FamilyMemberController extends Controller
 
         if ($request['family_code'] != null) {
             $findFamily =  DB::table('families')->where('family_code', $family_member->family_code)->first();
-            if($findFamily == null){
-                $family = new Family();
-                $family->family_code = $family_member->family_code;
-                $family->no_of_members = 1;
-                $family->save();
+            // if($findFamily == null){
+            //     $family = new Family();
+            //     $family->family_code = $family_member->family_code;
+            //     $family->no_of_members = 1;
+            //     $family->save();
 
-                $relief_recipient = new ReliefRecipient();
-                $relief_recipient->family_code = $family_member->family_code;
-                $relief_recipient->recipient_type = 'Non-Evacuee';
-                $relief_recipient->save();
-            }else if($findFamily != null){
+            //     $relief_recipient = new ReliefRecipient();
+            //     $relief_recipient->family_code = $family_member->family_code;
+            //     $relief_recipient->recipient_type = 'Non-Evacuee';
+            //     $relief_recipient->save();}else 
+                
+            if($findFamily != null){
                 $family = Family::find($findFamily->id);
                 $family->no_of_members = $family->no_of_members+1;
                 $family->save();
@@ -147,7 +148,7 @@ class FamilyMemberController extends Controller
             'gender'            => ['required', 'string', 'max:255', 'regex:/^[A-Za-z]+$/'],
             'birthdate'         => ['required', 'date_format:Y-m-d'],
             'address'           => ['required', 'string', 'max:255', 'alpha_spaces'],
-            'family_code'       => ['nullable', 'string', 'max:255'],
+            'family_code'       => ['nullable', 'string', 'max:255', 'exists:families,family_code'],
             'is_family_head'    => ['required', 'string', 'max:255'],
             'sectoral_classification' => ['required', 'string', 'max:255', 'alpha_spaces'],
         ]);
@@ -166,35 +167,38 @@ class FamilyMemberController extends Controller
 
         if ($request['family_code'] != null) {
             $findFamily =  DB::table('families')->where('family_code', $family_member->family_code)->first();
-            if($findFamily == null){
-                $family = new Family();
-                $family->family_code = $family_member->family_code;
-                $family->no_of_members = 1;
-                $family->save();
+            // if($findFamily == null){
+            //     $family = new Family();
+            //     $family->family_code = $family_member->family_code;
+            //     $family->no_of_members = 1;
+            //     $family->save();
 
-                $relief_recipient = new ReliefRecipient();
-                $relief_recipient->family_code = $family_member->family_code;
-                $relief_recipient->recipient_type = 'Non-Evacuee';
-                $relief_recipient->save();
-            }else if($findFamily != null){
+            //     $relief_recipient = new ReliefRecipient();
+            //     $relief_recipient->family_code = $family_member->family_code;
+            //     $relief_recipient->recipient_type = 'Non-Evacuee';
+            //     $relief_recipient->save();}else 
+                
+            if($findFamily != null){
                 $family = Family::find($findFamily->id);
                 $family->no_of_members = $family->no_of_members+1;
                 $family->save();
             }
+        }else if($request['family_code'] == null){
+            if($prev_family_member_family_code != null){
+                $findFamily =  DB::table('families')->where('family_code', $prev_family_member_family_code)->first();
+                if ($findFamily->no_of_members == 1) {
+                    $family = Family::find($findFamily->id);
+                    $family->delete();
+                    // $relief_recipient = DB::table('relief_recipients')->where('family_code', $prev_family_member_family_code);
+                    // $relief_recipient->delete();
+                } else {
+                    $family = Family::find($findFamily->id);
+                    $family->no_of_members = $family->no_of_members-1;
+                    $family->save();
+                }
+            } 
         }
-        if($prev_family_member_family_code != null){
-            $findFamily =  DB::table('families')->where('family_code', $prev_family_member_family_code)->first();
-            if ($findFamily->no_of_members == 1) {
-                $family = Family::find($findFamily->id);
-                $family->delete();
-                $relief_recipient = DB::table('relief_recipients')->where('family_code', $prev_family_member_family_code);
-                $relief_recipient->delete();
-            } else {
-                $family = Family::find($findFamily->id);
-                $family->no_of_members = $family->no_of_members-1;
-                $family->save();
-            }
-        }
+        
         $request->session()->flash('message', 'Resident updated successfully!');
         return redirect()->route('residents.index');
     }
@@ -211,13 +215,15 @@ class FamilyMemberController extends Controller
         if ($family_member) {
             $findFamily =  DB::table('families')->where('family_code', $family_member->family_code)->first();
         //   $count_members = DB::table('families')->where('family_code', $family_member->family_code)->count();
-            if ($findFamily->no_of_members == 1) {
+            if ($findFamily == null) {
+                $family_member->delete();
+            }else if ($findFamily->no_of_members == 1) {
                     $family_member->delete();
                     $family = Family::find($findFamily->id);
                     $family->delete();
-                    $relief_recipient = DB::table('relief_recipients')->where('family_code', $family_member->family_code);
-                    $relief_recipient->delete();
-            } else {
+                    // $relief_recipient = DB::table('relief_recipients')->where('family_code', $family_member->family_code);
+                    // $relief_recipient->delete();
+            } else if ($findFamily->no_of_members > 1) {
                     $family_member->delete();
                     $family = Family::find($findFamily->id);
                     $family->no_of_members = $family->no_of_members-1;
@@ -239,7 +245,7 @@ class FamilyMemberController extends Controller
         //     ->get();
         // dd($checkedResidents);
 
-        $family_members = DB::table('family_members')->select('id', 'name', 'sectoral_classification')->get();
+        $family_members = DB::table('family_members')->where('family_code', null)->select('id', 'name', 'sectoral_classification')->get();
 
         //$family_members = FamilyMember::all();
         return view('admin.relief-recipients-resource.groupFamilyMembers', ['family_members' => $family_members]);
@@ -247,33 +253,36 @@ class FamilyMemberController extends Controller
 
     public function groupResidents(Request $request)
     {
-
+        
         $validated = $request->validate([
-            'address'              => ['required', 'string', 'max:255']
+            'selectedResidents'              => ['required']
         ]);
-
+        //dd($request->selectedResidents);
         $family_code = 'eLIKAS-' . Str::random(6);
-        $no_of_members = 1;
-        if ($request->selectedResidents) {
-            foreach ($request->selectedResidents as $selectedResident) {
-                $no_of_members = +1;
-                $family_member = FamilyMember::find($selectedResident);
-                $family_member->family_code   = $family_code;
-                $family_member->save();
-            }
+
+        $family = new Family();
+        $family->family_code = $family_code;
+        $family->no_of_members = count($request->selectedResidents);
+        $family->save();
+        
+        foreach ($request->selectedResidents as $selectedResident) {
+            $family_member = FamilyMember::find($selectedResident);
+            $family_member->family_code   = $family_code;
+            $family_member->save();
         }
 
 
-        $family_member_rep = FamilyMember::find($request->selectedRepresentative);
-        $family_member_rep->is_representative = 'Yes';
-        $family_member_rep->save();
+        // $family_member_rep = FamilyMember::find($request->selectedRepresentative);
+        // $family_member_rep->is_representative = 'Yes';
+        
+        // $family_member_rep->save();
 
-        $relief_recipient = new ReliefRecipient();
-        $relief_recipient->family_code     = $family_code;
-        $relief_recipient->no_of_members     = $no_of_members;
-        $relief_recipient->address     = $validated['address'];
-        $relief_recipient->recipient_type     = 'Non-Evacuee';
-        $relief_recipient->save();
+        // $relief_recipient = new ReliefRecipient();
+        // $relief_recipient->family_code     = $family_code;
+        // $relief_recipient->no_of_members     = count($request->checkedResidents);
+        // $relief_recipient->address     = $validated['address'];
+        // $relief_recipient->recipient_type     = 'Non-Evacuee';
+        // $relief_recipient->save();
         $request->session()->flash('message', 'Group Resident successfully!');
         return redirect()->route('residents.index');
     }
